Improve formatting of code
