+++ ast/node.go
@@ -3,8 +3,6 @@ package ast
 import (
 	"errors"
 	"fmt"
-	"strconv"
-	"strings"
 
 	"github.com/NeowayLabs/nash/token"
 )
@@ -380,39 +378,6 @@ func (l *BlockNode) Push(n Node) {
 	l.Nodes = append(l.Nodes, n)
 }
 
-func (l *BlockNode) String() string {
-	nodes := l.Nodes
-	content := make([]string, 0, 8192)
-
-	for i := 0; i < len(nodes); i++ {
-		addEOL := false
-		node := nodes[i]
-
-		nodebytes := node.String()
-
-		if i == 0 && node.Type() == NodeComment && strings.HasPrefix(node.String(), "#!") {
-			nodebytes += "\n"
-		}
-
-		if (node.Type() == NodeAssignment) && i < (len(nodes)-1) {
-			nextNode := nodes[i+1]
-
-			switch nextNode.Type() {
-			case NodeComment, NodeFnDecl:
-				addEOL = true
-			}
-		}
-
-		if addEOL {
-			nodebytes += "\n"
-		}
-
-		content = append(content, nodebytes)
-	}
-
-	return strings.Join(content, "\n")
-}
-
 // IsEqual returns if it is equal to the other node.
 func (l *BlockNode) IsEqual(other Node) bool {
 	if l == other {
@@ -480,11 +445,6 @@ func (n *ImportNode) IsEqual(other Node) bool {
 	return false
 }
 
-// String returns the string representation of the import
-func (n *ImportNode) String() string {
-	return `import ` + n.path.String()
-}
-
 // NewSetenvNode creates a new assignment node
 func NewSetenvNode(info token.FileInfo, name string) *SetenvNode {
 	return &SetenvNode{
@@ -516,11 +476,6 @@ func (n *SetenvNode) IsEqual(other Node) bool {
 	}
 
 	return n.varName == o.varName
-}
-
-// String returns the string representation of assignment
-func (n *SetenvNode) String() string {
-	return "setenv " + n.varName
 }
 
 // NewAssignmentNode creates a new assignment
@@ -583,17 +538,6 @@ func (n *AssignmentNode) IsEqual(other Node) bool {
 	}
 
 	return false
-}
-
-// String returns the string representation of assignment statement
-func (n *AssignmentNode) String() string {
-	obj := n.val
-
-	if obj.Type().IsExpr() {
-		return n.name + " = " + obj.String()
-	}
-
-	return "<unknown>"
 }
 
 // NewExecAssignNode creates a new node for executing something and store the
@@ -665,11 +609,6 @@ func (n *ExecAssignNode) IsEqual(other Node) bool {
 	}
 
 	return false
-}
-
-// String returns the string representation of command assignment statement
-func (n *ExecAssignNode) String() string {
-	return n.name + " <= " + n.cmd.String()
 }
 
 // NewCommandNode creates a new node for commands
@@ -759,95 +698,6 @@ func (n *CommandNode) IsEqual(other Node) bool {
 	return n.name == o.name
 }
 
-func (n *CommandNode) toStringParts() ([]string, int) {
-	var (
-		content  []string
-		line     string
-		last     = len(n.args) - 1
-		totalLen = 0
-	)
-
-	for i := 0; i < len(n.args); i += 2 {
-		var next string
-
-		arg := n.args[i].String()
-
-		if i < last {
-			next = n.args[i+1].String()
-		}
-
-		if i == 0 {
-			arg = n.name + " " + arg
-		}
-
-		if arg[0] == '-' {
-			if line != "" {
-				content = append(content, line)
-				line = ""
-			}
-
-			if next[0] != '-' {
-				line += arg + " " + next
-			} else {
-				content = append(content, arg, next)
-			}
-		} else if next != "" {
-			line += arg + " " + next
-		} else {
-			line += arg
-		}
-
-		totalLen += len(arg) + len(next) + 1
-
-	}
-
-	if line != "" {
-		content = append(content, line)
-	}
-
-	return content, totalLen
-}
-
-func (n *CommandNode) multiString() string {
-	content, totalLen := n.toStringParts()
-
-	if totalLen < 50 {
-		return "(" + strings.Join(content, " ") + ")"
-	}
-
-	content[0] = "\t" + content[0]
-
-	gentab := func(n int) string { return strings.Repeat("\t", n) }
-	tabLen := (len(content[0]) + 7) / 8
-
-	for i := 1; i < len(content); i++ {
-		content[i] = gentab(tabLen) + content[i]
-	}
-
-	return "(\n" + strings.Join(content, "\n") + "\n)"
-}
-
-// String returns the string representation of command statement
-func (n *CommandNode) String() string {
-	if n.multi {
-		return n.multiString()
-	}
-
-	var content []string
-
-	content = append(content, n.name)
-
-	for i := 0; i < len(n.args); i++ {
-		content = append(content, n.args[i].String())
-	}
-
-	for i := 0; i < len(n.redirs); i++ {
-		content = append(content, n.redirs[i].String())
-	}
-
-	return strings.Join(content, " ")
-}
-
 // NewPipeNode creates a new command pipeline
 func NewPipeNode(info token.FileInfo, multi bool) *PipeNode {
 	return &PipeNode{
@@ -897,83 +747,6 @@ func (n *PipeNode) IsEqual(other Node) bool {
 	}
 
 	return cmpInfo(n, other)
-}
-
-func (n *PipeNode) multiString() string {
-	totalLen := 0
-
-	type cmdData struct {
-		content  []string
-		totalLen int
-	}
-
-	content := make([]cmdData, len(n.cmds))
-
-	for i := 0; i < len(n.cmds); i++ {
-		cmdContent, cmdLen := n.cmds[i].toStringParts()
-
-		content[i] = cmdData{
-			cmdContent,
-			cmdLen,
-		}
-
-		totalLen += cmdLen
-	}
-
-	if totalLen+3 < 50 {
-		result := "("
-
-		for i := 0; i < len(content); i++ {
-			result += strings.Join(content[i].content, " ")
-
-			if i < len(content)-1 {
-				result += " | "
-			}
-		}
-
-		return result + ")"
-	}
-
-	gentab := func(n int) string { return strings.Repeat("\t", n) }
-
-	result := "(\n"
-
-	for i := 0; i < len(content); i++ {
-		cmdContent := content[i].content
-		cmdContent[0] = "\t" + cmdContent[0]
-		tabLen := (len(cmdContent[0]) + 7) / 8
-
-		for j := 1; j < len(cmdContent); j++ {
-			cmdContent[j] = gentab(tabLen) + cmdContent[j]
-		}
-
-		result += strings.Join(cmdContent, "\n")
-
-		if i < len(content)-1 {
-			result += " |\n"
-		}
-	}
-
-	return result + "\n)"
-}
-
-// String returns the string representation of pipeline statement
-func (n *PipeNode) String() string {
-	if n.multi {
-		return n.multiString()
-	}
-
-	ret := ""
-
-	for i := 0; i < len(n.cmds); i++ {
-		ret += n.cmds[i].String()
-
-		if i < (len(n.cmds) - 1) {
-			ret += " | "
-		}
-	}
-
-	return ret
 }
 
 // NewRedirectNode creates a new redirection node for commands
@@ -1039,33 +812,6 @@ func (r *RedirectNode) IsEqual(other Node) bool {
 	return false
 }
 
-// String returns the string representation of redirect
-func (r *RedirectNode) String() string {
-	var result string
-
-	if r.rmap.lfd == r.rmap.rfd {
-		if r.location != nil {
-			return "> " + r.location.String()
-		}
-
-		return ""
-	}
-
-	if r.rmap.rfd >= 0 {
-		result = ">[" + strconv.Itoa(r.rmap.lfd) + "=" + strconv.Itoa(r.rmap.rfd) + "]"
-	} else if r.rmap.rfd == RedirMapNoValue {
-		result = ">[" + strconv.Itoa(r.rmap.lfd) + "]"
-	} else if r.rmap.rfd == RedirMapSupress {
-		result = ">[" + strconv.Itoa(r.rmap.lfd) + "=]"
-	}
-
-	if r.location != nil {
-		result = result + " " + r.location.String()
-	}
-
-	return result
-}
-
 // NewRforkNode creates a new node for rfork
 func NewRforkNode(info token.FileInfo) *RforkNode {
 	return &RforkNode{
@@ -1116,26 +862,6 @@ func (n *RforkNode) IsEqual(other Node) bool {
 	return n.tree.IsEqual(o.tree)
 }
 
-// String returns the string representation of rfork statement
-func (n *RforkNode) String() string {
-	rforkstr := "rfork " + n.arg.String()
-	tree := n.Tree()
-
-	if tree != nil {
-		rforkstr += " {\n"
-		block := tree.String()
-		stmts := strings.Split(block, "\n")
-
-		for i := 0; i < len(stmts); i++ {
-			stmts[i] = "\t" + stmts[i]
-		}
-
-		rforkstr += strings.Join(stmts, "\n") + "\n}"
-	}
-
-	return rforkstr
-}
-
 // NewCommentNode creates a new node for comments
 func NewCommentNode(info token.FileInfo, val string) *CommentNode {
 	return &CommentNode{
@@ -1166,11 +892,6 @@ func (n *CommentNode) IsEqual(other Node) bool {
 	}
 
 	return n.val == o.val
-}
-
-// String returns the string representation of comment
-func (n *CommentNode) String() string {
-	return n.val
 }
 
 // NewIfNode creates a new if block statement
@@ -1286,56 +1007,6 @@ func (n *IfNode) IsEqual(other Node) bool {
 	return expectedTree.IsEqual(valueTree)
 }
 
-// String returns the string representation of if statement
-func (n *IfNode) String() string {
-	var lstr, rstr string
-
-	lstr = n.lvalue.String()
-	rstr = n.rvalue.String()
-
-	ifStr := "if " + lstr + " " + n.op + " " + rstr + " {\n"
-
-	ifTree := n.IfTree()
-
-	block := ifTree.String()
-	stmts := strings.Split(block, "\n")
-
-	if strings.TrimSpace(block) != "" {
-		for i := 0; i < len(stmts); i++ {
-			stmts[i] = "\t" + stmts[i]
-		}
-	}
-
-	ifStr += strings.Join(stmts, "\n") + "\n}"
-
-	elseTree := n.ElseTree()
-
-	if elseTree != nil {
-		ifStr += " else "
-
-		elseBlock := elseTree.String()
-		elsestmts := strings.Split(elseBlock, "\n")
-
-		for i := 0; i < len(elsestmts); i++ {
-			if !n.IsElseIf() {
-				elsestmts[i] = "\t" + elsestmts[i]
-			}
-		}
-
-		if !n.IsElseIf() {
-			ifStr += "{\n"
-		}
-
-		ifStr += strings.Join(elsestmts, "\n")
-
-		if !n.IsElseIf() {
-			ifStr += "\n}"
-		}
-	}
-
-	return ifStr
-}
-
 // NewFnDeclNode creates a new function declaration
 func NewFnDeclNode(info token.FileInfo, name string) *FnDeclNode {
 	return &FnDeclNode{
@@ -1400,39 +1071,6 @@ func (n *FnDeclNode) IsEqual(other Node) bool {
 	return cmpInfo(n, other)
 }
 
-// String returns the string representation of function declaration
-func (n *FnDeclNode) String() string {
-	fnStr := "fn"
-
-	if n.name != "" {
-		fnStr += " " + n.name + "("
-	}
-
-	for i := 0; i < len(n.args); i++ {
-		fnStr += n.args[i]
-
-		if i < (len(n.args) - 1) {
-			fnStr += ", "
-		}
-	}
-
-	fnStr += ") {\n"
-
-	tree := n.Tree()
-
-	stmts := strings.Split(tree.String(), "\n")
-
-	for i := 0; i < len(stmts); i++ {
-		if len(stmts[i]) > 0 {
-			fnStr += "\t" + stmts[i] + "\n"
-		}
-	}
-
-	fnStr += "}\n"
-
-	return fnStr
-}
-
 // NewFnInvNode creates a new function invocation
 func NewFnInvNode(info token.FileInfo, name string) *FnInvNode {
 	return &FnInvNode{
@@ -1486,23 +1124,6 @@ func (n *FnInvNode) IsEqual(other Node) bool {
 	return cmpInfo(n, other)
 }
 
-// String returns the string representation of function invocation
-func (n *FnInvNode) String() string {
-	fnInvStr := n.name + "("
-
-	for i := 0; i < len(n.args); i++ {
-		fnInvStr += n.args[i].String()
-
-		if i < (len(n.args) - 1) {
-			fnInvStr += ", "
-		}
-	}
-
-	fnInvStr += ")"
-
-	return fnInvStr
-}
-
 // NewBindFnNode creates a new bindfn statement
 func NewBindFnNode(info token.FileInfo, name, cmd string) *BindFnNode {
 	return &BindFnNode{
@@ -1536,11 +1157,6 @@ func (n *BindFnNode) IsEqual(other Node) bool {
 	}
 
 	return n.name == o.name && n.cmdname == o.cmdname
-}
-
-// String returns the string representation of bindfn
-func (n *BindFnNode) String() string {
-	return "bindfn " + n.name + " " + n.cmdname
 }
 
 // NewDumpNode creates a new dump statement
@@ -1588,15 +1204,6 @@ func (n *DumpNode) IsEqual(other Node) bool {
 	return false
 }
 
-// String returns the string representation of dump node
-func (n *DumpNode) String() string {
-	if n.filename != nil {
-		return "dump " + n.filename.String()
-	}
-
-	return "dump"
-}
-
 // NewReturnNode create a return statement
 func NewReturnNode(info token.FileInfo) *ReturnNode {
 	return &ReturnNode{
@@ -1641,15 +1248,6 @@ func (n *ReturnNode) IsEqual(other Node) bool {
 	}
 
 	return false
-}
-
-// String returns the string representation of return statement
-func (n *ReturnNode) String() string {
-	if n.arg != nil {
-		return "return " + n.arg.String()
-	}
-
-	return "return"
 }
 
 // NewForNode create a new for statement
@@ -1703,25 +1301,6 @@ func (n *ForNode) IsEqual(other Node) bool {
 
 	return n.identifier == o.identifier &&
 		n.inVar == o.inVar
-}
-
-// String returns the string representation of for statement
-func (n *ForNode) String() string {
-	ret := "for"
-
-	if n.identifier != "" {
-		ret += " " + n.identifier + " in " + n.inVar
-	}
-
-	ret += " {\n"
-
-	if n.tree != nil {
-		ret += n.tree.String() + "\n"
-	}
-
-	ret += "}"
-
-	return ret
 }
 
 func cmpInfo(n, other Node) bool {
